requirements.txt
Flask==2.3.3
flask-socketio==5.3.5
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
WTForms==3.1.1
sqlalchemy==2.0.44
python-dotenv==1.0.0
requests==2.31.0
pytz==2023.3
python-barcode==0.15.1
Pillow==10.0.1
qrcode==7.4.2
openai==2.8.1
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0







psycopg2-binary
pandas==2.1.0
openpyxl==3.1.2
gunicorn==21.2.0
eventlet==0.33.3
